requirements.txt
python-dotenv
pandas
numpy
psycopg2-binary
certifi 
setuptools
scikit-learn
#-e .# refering to setup.py